requirements.txt
opencv-python
ultralytics
clip-interrogator

torch
torchsde
torchvision
einops
transformers>=4.25.1
safetensors>=0.3.0
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil

#non essential dependencies:
kornia>=0.7.1
spandrel

#comfyui
importlib_metadata
huggingface_hub
opencv-python-headless>=4.7.0.72
filelock
numpy
scikit-image
python-dateutil
mediapipe
svglib
fvcore
yapf
omegaconf
ftfy
addict
yacs
trimesh[easy]
albumentations
scikit-learn
